backend/requirements.txt
django==5.1.7
django-jazzmin==3.0.1
fastapi[standard]==0.115.11
pillow==11.1.0
python-dotenv==1.0.0
pymysql==1.1.1
django-unfold==0.53.0
django-import-export==4.3.8.dev3